pyproject.toml
[tool.poetry]
name = "tesla-alert-bot"
version = "0.1.0"
description = "A real-time Tesla news and price alert bot using Telegram."
authors = ["Jinsoo Heo <dev.koriel@gmail.com>"]
license = "MIT"
readme = "README.md"
homepage = "https://github.com/devkoriel/tesla_alert_bot"
repository = "https://github.com/devkoriel/tesla_alert_bot"
documentation = "https://github.com/devkoriel/tesla_alert_bot"
package-mode = false

[tool.poetry.dependencies]
python = ">=3.9,<4.0"
python-telegram-bot = { version = "==22.0", extras = ["job-queue"] }
requests = "==2.32.3"
beautifulsoup4 = "==4.13.3"
schedule = "==1.2.2"
openai = "^1.68.0"
psycopg2-binary = "^2.9.10"
redis = "^5.2.1"
feedparser = "^6.0.11"
python-dotenv = "^1.0.1"
nest_asyncio = "^1.6.0"

[tool.poetry.group.dev.dependencies]
black = "^25.1.0"
isort = "^6.0.1"
autoflake = "^2.3.1"
autopep8 = "^2.3.2"
flake8 = "^7.1.2"
pre-commit = "^3.3.3"

[build-system]
requires = ["poetry-core>=1.5.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 120
target-version = ["py39"]

[tool.isort]
profile = "black"
line_length = 120

[tool.flake8]
max-line-length = 120
exclude = [".git", "__pycache__", "venv"]

[tool.autopep8]
max_line_length = 120
